address review feedback
- move FreeType and Skrifa specific code into own modules
- move compare_glyphs() function from bin to lib
- add tests for RegularizingPen cases
