+++ fauntlet/src/lib.rs
@@ -1,5 +1,7 @@
+mod compare_glyphs;
 mod font;
 mod pen;
 
+pub use compare_glyphs::compare_glyphs;
 pub use font::{Font, FreeTypeInstance, InstanceOptions, SharedFontData, SkrifaInstance};
 pub use pen::{RecordingPen, RegularizingPen};
